move SpacyConfig to configs folder
